refactor: 공통 헤더와 페이지 헤더 분리
